Move AuthGate to pkg/auth (#1327)
Moving forward, it will be needed to implement Helm 3 support (see
#1309), and it was in an internal package, so it couldn't be imported
outside the `cmd/tiller-proxy` directory.
